Ensure presentation compiler futures get completed (#131)
Previously, errors were uncaught in the presentation compiler. This
meant that Metals would forever block if you called `.get()` on mtags
methods without a timeout. This also made it difficult to troubleshoot
an issue since the stack trace would be logged by the execution context
instead of by the mtags client.

Now, we catch errors and fail the future so that the consumer of mtags
can handle the error appropriately. This PR adds a new test case that
previously timed out but now you can instead gracefully handle the
thrown error.

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>