Java Header Compiler Plugin (#133)
This plugin lets the "prune compiler" stop crawling the codebase when
using the `-sourcepath` option. For example, if you want to implement
completions for a file `A.java` and it depends on `B.java`, which
transitively depends on `C.java`, then you would have to typecheck the
full transitive graph if you run javac without this header compiler.
With the header compiler, we remove the bodies of `B.java` so that we
only typecheck A and B instead.

This PR is a no-op because it's not used yet in the codebase.

---------

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>