Java: Go to Definition (#136)
This PR implement "Go to Definiton" and "Go to Type Definition" using
the new Java prune compiler infrastructure. The implementation diverges
from Metals idioms by resolving from the -sourcepath as much as
possible.

Historically, Metals has heavily relied on the -classpath via BSP and
then queried its own "definition index" to locate 1st-party and
3rd-party definitions. However, now that we use the -sourcepath we can
stop using the definition index for 1st-party dependencies.

Sidenote: it's not implemented in this PR, but we can stop using the
definition index for 3rd-party dependencies as well and once we
implement that, we can heavily cut down on indexing times and memory
usage. Very exciting!

---------

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>