Retire FOLLOW_SYMLINK feature flag (#145)
My plan was to use this feature flag as a test bed to validate impact of
this exact change. However, due to a bug in how we tested feature flags,
this was enabled at 100% rollout from the start so I think it's best to
retire the flag instead of regressing people's experience.