[PLAT-146135]: remove mtags3 project to unblock dogfooding (#152)
When dogfooding our internal Metals in the Metals repo, the logs get a
lot of errors from trying to resolve mtags_3.3.3. This happens because
it's still a project in the build, and Metals runs the Doctor
proactively to identify issues in the workspace. One of the issues it
correctly detects is that mtags_3.3.3 doesn't exist. By removing this
project, it should stop trying to resolve mtags_3.3.3 and the logs will
be less noisy.