[PLAT-146195]: ignore noisy report when using goto-definition for local Java variables (#154)
In mtags-java, we have a flawed way to compute local variables from the
old completion/hover providers. I have not yet fixed this because
everything is mostly working as I expect it anyways, but I noticed that
we're creating compiler bug reports when to do goto-definition in local
variables
```
empty definition using pc, found symbol in pc: localUNKNOWN_BECAUSE_WE_HAVENT_HANDLED_LOCALS_YET%
```
We can ignore these errors so this PR just stops Metals from creating
the reports when this happens.

## Test Plan

This PR adds a test case that reproduced the old problem. I manually
confirmed it created a bug report before the fix and it doesn't do it
after the fix.