[PLAT-146199]: fix spurious errors when doing repo-wide symbol search (#156)
Previously, repo-wide symbol search had spurious crashes because it was
not thread-safe. Metals would just not return meaningful results for
valid queries. This has never been a problem before because the old
BSP-based symbol search was single-threaded. However, the new mbt-based
symbol search is multi-threaded and calls `visitor.onMatch` in parallel.

Now, the search visitor is thread-safe so that I'm unable to reproduce
the errors I saw in the logs before.