Java: Selection Range (#143)
This implements `textDocument/selectionRange` in LSP. This feature
allows users to expand/collapse their text selection based on the AST of
the code. The default implementation in VS Code (no LSP) tends to expand
too quickly so that it's difficult to, for example, select a single
argument inside a method invocation.

**With Selection Range**


https://github.com/user-attachments/assets/4c882f71-292c-4a38-8ae7-34e0613ccf59

**No Selection Range**


https://github.com/user-attachments/assets/cd9391b9-1920-45f3-ad38-6eb88c0e3057

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>