[PLAT-146122]: refactor - use Java bindings for SemanticDB `Language` (#161)
This is the first step towards making the mbt symbol index support
transpiling Protobuf into Java code on the fly. The Scala bindings for
SemanticDB are from the upstream org.scalameta:semanticdb dependency
that is complicated to update. Instead, we're slowly moving over to the
equivalent Java bindings for SemanticDB, which we generate in the Metals
build.

This PR is a no-op, we're just replacing the Scala bindings with the
Java ones.