Less flaky CI failures (#178)
I've noticed that the Bazel CI job tends to fail with what seems to be a
race condition in the build. The `testOnly` command has a custom
dependency on `publishLocal` steps that is likely the root cause, and I
see that the sbt CI job did a separate `sbt quick-publish-local` step
that might help here.