Speed up indexing by moving to classpath-based indexing (#174)
Our metrics show that indexing dependency sources is the second-slowest
part of the Metals indexing pipeline taking upwards of 14s in the p95
case, and 40s in p99. This overhead comes from the running a "toplevel
indexer" on all the `*-sources.jar` files of transitive dependencies.
While this toplevel indexer is fast (it processes ~1m loc/s) and we
cache the result of this indexing step, it still contributes to
everything feeling sluggish.

Now, we have a new `"definitionIndexStrategy":"classpath"` configuration
option that skips the toplevel indexing part altogether. Instead, we
build a much smaller index over what *packages* are defined by which jar
file. For example, the jar file `scala-library.jar` maps to
`Set("scala/", "scala/collections/", "scala/util/", ...)`. Using this
index, we can resolve "go to definition" requests at query time by
reading the `source` debug attribute in the classfile that defines the
query symbol, replacing our previous usage of the toplevels index.

## Test Strategy

This PR enables the new `"classpath"` indexing strategy by default in
tests. A bunch of tests fail if the classpath-based indexer does
nothing, and almost all tests pass when it's working correctly. The
failing tests fall in ~two categories:

- "Find implementations" no longer works for symbols from 3rd-party
dependencies. For example, "Find implementations" on `Set` from
scala-library won't show all the available set classes in the stdlib.
This is fixable by indexing classfiles, which is more accurate and
faster than the source-based approach anyways. Either way, it's not a
critical feature when "Find implementations" doesn't work for 1st-party
dependencies.
- "Go to symbol in workspace" for symbolic names like `::` in external
dependencies. Also not an important use-case, and solvable by decoding
`$plus` symbols in classfiles in classpath indexing (how it should have
been implemented in the first place).

## Manual Benchmark

I ran `ManualSuite` with and without the new classpath-based indexing in
universe where I have synced
`elastic-spark/lakeflow/assets/list/impl/InMemorySortingAssetFetcher.scala`
and a few other files.

- Old `sources`-based indexing (cold, no cache): ~1m5s
- Old `sources`-based indexing (hot, cached): ~4s.
- New `classpath`-based indexing (cold and hot, there is no cache): ~1s

To put this in perspective, with a large number of synced targets, the
total runtime of starting Metals and doing a single "Go to definition"
request goes from ~93s (old+cold) or ~32s (old+hot), down to ~28s
(cold).

On memory footprint, the difference is smaller than I was expecting. The
definition index went down from ~800mb to ~500mb. I recall seeing that
the definition index was ~2-3gb at some point, and that might still be
true if I sync a different set of targets. Either way, the memory
footprint of this new solution should not grow much even if you have a
gigantic set of external deps (for example, everything in universe is
likely fine).