[PLAT-146334] Java Formatting & Organize Imports via google-java-format (#185)
Previously, we only supported formatting Java code with the Eclipse
formatter. Now, we additionally support formatting with
google-java-format.

We get "Organize Imports" for free from google-java-format since it also
supports reordering imports and removing unused imports. Unfortunately,
it's not as easy to do the same for the Scala language due to implicits.