Apply consistent backpressure to `didCreate` and `didChange` events (#189)
Previously, Metals used a mix of solutions to avoid duplicate processing
of events for file creation and file edits. For example, we debounced
diagnostics but not other parts of the indexing pipeline. This was
problematic because it meant that subtle bugs could cause Metals to miss
important file watching events resulting in a stale index, which
eventually results in a buggy user experience

Now, Metals uses the same `BatchedFunction.backpressure` helper to
apply a debounce-like effect on file creation/edit events. This means
we wait 500ms before processing these events, and batch the various
`textDocument/didSave`, `textDocument/didChange`, `textDocument/didOpen`
and `workspace/didChangeWatchedFiles` events that arrive through that
window. It's not a true debounce effect because we unconditionally flush
the events after the delay from the first invocation.