Format codebase with google-java-format (#202)
Previously, we didn't enforce any specific formatting for Java files  in
the codebase. I just upgraded to the latest Metals version and now I'm
formatting Java files on save with google-java-format, which is causing
large diffs in my PRs.

Now, we enforce google-java-format in CI so that you can use
format-on-save with Metals without creating large unrelated diffs.