[PLAT-147154]: Fix fatal javac error (#207)
This issue was reported [on
Friday](https://databricks.slack.com/archives/C09E2L39N8J/p1761926895993719)
and I didn't get a reliable reproduction until last night. The PR #191
was an attempt to address this issue but it did not fix the root bug.
This PR fixes the root bug and includes a minimized reproduction test
case to prevent it from happening again.

**Summary**: if you transitively imported a file without a toplevel
class then it would break **all** Java LSP features. This happened
because we delegated the method `super.inferBinaryName(..)` when we
didn't detect a toplevel class, and the compiler threw an
`IllegalArgumentException` taking down the whole compilation task.