Fix report path to be clickable (#211)
The compiler crash report path contains the target name in parenthesis,
but the VS Code link parser is not able to parse and link to these
paths. They appear in the logs, and it is a small quality of life
improvement to be able to just click-to-open such files.

Bazel targets contain `/` and these are turned into directories, which
is acceptable, but ( and : are problematic.