[PLAT-139354]: fix organize imports for Scala (#212)
Previously, running "Organize Imports" did not remove unused imports in
Scala. The root problem was that the PC SemanticDB provider did no
populate the `TextDocument.diagnostics` field. Now, organizing imports
works for Scala.

Context: Scalafix relies on diagnostics to know what imports are unused
because that's the only way to get 100% reliable analysis of what
imports are actually unused. It's really tricky to do the analysis
outside the compiler due to implicits.


https://github.com/user-attachments/assets/462559fc-343c-41e2-aa17-3cce29d46288