Use Java compiler for mtags toplevel indexing (#199)
Previously, we used two different Java outline indexers:

- A toplevel one based on a custom Java scanner, which was super fast
(~1.5m loc/s) but did not support indexing methods. We ran this against
all external `*-sources.jar` as well as all Java sources in the repo for
workspace/symbol search.
- Qdox-based, which was ~5x slower (~300k loc/s) but did support
methods. We only ran this one on individual files at query time during
"Go to definition" to resolve the definition location of methods (which
the toplevel indexer didn't index).

Now, we've added a third option: an outline indexer that uses javac
directly (similar to how google-java-format works). There's a new
feature flag to control whether to use the javac-based indexer or the
qdox-based indexer. The javac indexer fixes several bugs that ensure we
have 100% compatibility with the output of semanticdb-javac. Just
working on this PR helped me fix a few nice bugs in semanticdb-javac
itself.

The goal is to use this new javac indexer to eventually power "Find
references" as well by building an index of referenced identifiers. This
is essentially redoing the `IdentifierIndex` approach, but with slightly
higher precision (we exclude 100% local variable references) and as part
of the same indexing pipeline (mtags definitions + references from the
same parse).

## Performance Benchmarks

Performance-wise, the javac parser is ~33% faster than qdox but it's
still ~3x slower than the toplevel indexer. When you index references,
it's ~4x slower than the toplevel indexer or doing ~400k lines/s. It's
not as fast as I was hoping, but it's good enough for now to 1) replace
qdox and 2) unblock further experimenentation with find-references.

```
[info] MetalsBench.mtagsJavaParse                    ss    3  12.576 ± 0.081   s/op
[info] MetalsBench.mtagsJavacParse                   ss    3   9.571 ± 1.800   s/op
[info] MetalsBench.mtagsJavaToplevelParse            ss    3   2.706 ± 0.725   s/op
```

Interestingly, ~30% of the runtime comes from creating a new file
manager and closing it per input so there might be room for
optimizations here when batch-indexing.
```
[info] MetalsBench.mtagsJavacBaseline    ss    3  5.471 ± 0.254   s/op # reused file manager
[info] MetalsBench.mtagsJavacBaseline    ss    3  8.425 ± 0.294   s/op # one file manager per input
```