Remove unnecessary find-references fallback in goto-definition (#214)
Previously, Metals had a homegrown solution to automatically trigger
"Find references" when receiving a "Go to definition" request at the
definition location. This was added in ~2019
(https://github.com/scalameta/metals/pull/775) to support the same UX as
in IntelliJ, and it was necessary back then because VS Code didn't
implement this behavior. Now, VS Code (and Cursor) do this automatically
so we can remove this special case.

I discovered this while working on find references because the homegrown
solution has very poor performance, it effectively triggers three rounds
of goto-definition requests in the worst case, which all trigger a PC
compilation (at least for Java right now).

## Test Plan

Local repro with a build of this commit



https://github.com/user-attachments/assets/8aa72df7-7654-4c13-b279-0a4807caf579