Display an LSP progress bar during mbt-v2 startup (#216)
While working on find references, I found it difficult to know when
Metals was ready to handle LSP "Find references" requests. If I
triggered it before mbt-v2 indexing completed, then I only got partial
results. Most of the time, this isn't a problem because the index loads
in ~3 seconds when its cached (but our production metrics reveal it
might be slower for many people).

Also, as we move mbt-v2 over to using the mtags-{scalac,javac} indexers,
cold indexing is going to get quite a bit slower because we're indexing
more (tracking referenced identifiers as well. So this problem will
become more painful soon!

This PR adds a new LSP `$/progress` bar for the mbt-v2 indexing step so
it's easier for users to know when Metals is "ready".

## Test Plan

Local build, `rm -rf .metals/index.mbt`, and confirm it opens this in
the Cursor status bar


https://github.com/user-attachments/assets/3ce91d63-b2fe-4e6f-963c-cda9dbf729a2

The status bar doesn't appear for me when the index is hot.