Optimize mbt-v2 cold indexing times (#217)
Somewhere along the way, mbt-v2 cold indexing times regressed severely
because of how we keep the `documentsKeys` array up-to-date. It was
taking upwards of 5 minutes(!!) to index all of universe when the cache
was cold. I only noticed this because I added the progress bar in #216.

Now, we only update the keys after we finish indexing all files. This
takes the cold indexing time down to ~1min20s, which still isn't great
but it's better than 5 minutes.