Use correct URIs in IndexedDocument (#220)
Use correct URIs in IndexedDocument

Previously, if you created a `IndexedDocument` from a jar URI (element
inside an open jar file), then it would crash to run javac with it. Now,
we reuse the same helper method as we're using elsewhere to compile
jar URIs, which converts the URI into an absolute path.

<!-- jj-stack-begin -->
- → **#220** ←
<!-- jj-stack-end -->