Optimization: disable unused feature to digest the build (#236)
For a very large set of synced targets, this change makes the "Syncing
workspace" phase from ~30 seconds to ~15 seconds. If I have no synced
targets, it goes from ~20 seconds down to ~5 seconds.

Context: the "syncing workspace" phase does the following steps:

- Digest the build (now disabled by default)
- Connect to BSP server
- Import BSP data
- Index BSP data
- Digest the build again (now disabled by default)

We can disable this feature by default because our users are not relying
on "Do you want to re-import the build?" prompts. Instead, users are
expected to trigger syncs manually.

In OSS Metals, this feature mirrors the same feature in IntelliJ
where it prompts the user to re-import the build when files like
build.sbt/build.gradle change.

The functionality still exists behind an off-by-default config option,
and it's enabled by default in the tests to keep CI green.

<!-- jj-stack-begin -->
- #235
[[Commit](https://github.com/REDACTED_ORG/metals/pull/235/commits/f1b13cd6d4444bbaad15228e79e92b9c7ca3ad43)]
- → **#236** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/236/commits/2580e09fb98eda36844cfe22e7db8239b930c33c)]
<!-- jj-stack-end -->