Stop force-enabling interactive SemanticDBs for Java (#230)
Previously, we treated Java files the same as we did standalone
script-like files (sbt, worksheets, mill builds, ...) by enabling
"interactive semanticdbs". While this enabled some nice functionality
like code lenses, it also results in sending additional requests to the
Java PC to build SemanticDB payloads, which the Java PC currently is
struggling to handle.

For now, let's disable this functionality since it's a nice-to-have and
it's causing problems for P0 issues like responsive diagnostics. We
will need a different approach to get code lenses working and it should
be work consistently for Java and Scala when we do that.

<!-- jj-stack-begin -->
- #232
[[Commit](https://github.com/REDACTED_ORG/metals/pull/232/commits/55dd8f599dba0b8640c55cd5689ee106cc135746)]
- → **#230** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/230/commits/0d6adace7e0a5bfcbedf0bd0173bfe9317620d06)]
- #227
[[Commit](https://github.com/REDACTED_ORG/metals/pull/227/commits/fe91f7ec624ea15113a60a732adbacf12ced49e8)]
- #210
[[Commit](https://github.com/REDACTED_ORG/metals/pull/210/commits/2ab179854f49f04eba6daf49f9e9242f6ae7eef2)]
- #231
[[Commit](https://github.com/REDACTED_ORG/metals/pull/231/commits/f1e5423dfa5c8c614a78fa0b9dbe439853a510a5)]
<!-- jj-stack-end -->