Optimization: support batch indexing API for the presentation compiler (#221)
Previously, when using the presentation compiler (PC), it was only
possible to SemanticDB index a single document at a time. Now, with the
-sourcepath option, this becomes expensive when you need to process
multiple times because it means you re-compute the sourcepath for every
requsted file (at least for Java).

Now, there's a new `PresentationCompiler.batchSemanticdbTextDocuments`
API in mtags-interfaces that solves this problem. For Scala, we can
trivially fallback to the single-document mode, which is functionally
correct but inefficient. For Java, this makes a huge different in UX for
the upcoming "Find references". In fact, this optimization is the
difference between getting 200+ results from 50+ files in ~3s instead of
getting 10 results in ~15s.

<!-- jj-stack-begin -->
- → **#221** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/221/commits/8d626c604d9b723621c6ded7f84651a0f3d501b1)]
- #224
[[Commit](https://github.com/REDACTED_ORG/metals/pull/224/commits/1a85024873a4aa8cfb07b83efb92675d1d18fac9)]
- #233
[[Commit](https://github.com/REDACTED_ORG/metals/pull/233/commits/ae259261b2fbf8c420c22a62c50f14172c4a3600)]
- #227
[[Commit](https://github.com/REDACTED_ORG/metals/pull/227/commits/5a9b98f42cef53213c56a913ad05b3ff00ee7259)]
- #238
[[Commit](https://github.com/REDACTED_ORG/metals/pull/238/commits/a73843f8cbc9c8552ce814a7aa65133150858f32)]
- #239
[[Commit](https://github.com/REDACTED_ORG/metals/pull/239/commits/972711c8ddb4518e37b144e206813a6904158741)]
- #210
[[Commit](https://github.com/REDACTED_ORG/metals/pull/210/commits/1a138facbec4e6c3a941ee10e5d7f43f145de128)]
<!-- jj-stack-end -->