Optimization: Cache javac compile tasks (#224)
Previously, every Java LSP request triggered a fresh javac analyze
phase, which was slow. Now, if the inputs have not changed (compile
options or source contents), then we reuse the previous analysis
resulting in big performance improvements across a burst of read-only
requests. For example, when the user is only navigating a file, we will
now run the analysis phase once for semantic highlighting, and then
reuse that analysis for all subsequent
goto-definition/find-references/parameter-hints/... requests.

## Test Plan

See the added benchmarks, which confirm that repeated requests with the
same inputs are indeed cached
```
[info] Benchmark                                            (repo)  Mode  Cnt   Score   Error  Units
[info] JavaPruneCompilerBench.javacPruneCacheTask2           trino    ss       10.118           s/op
[info] JavaPruneCompilerBench.javacPruneCacheTask3           trino    ss       10.668           s/op
[info] JavaPruneCompilerBench.javacPruneCacheTaskBaseline1   trino    ss       10.000           s/op
[info] JavaPruneCompilerBench.javacPruneCacheTaskBaseline2   trino    ss       18.687           s/op
[info] JavaPruneCompilerBench.javacPruneCacheTaskBaseline3   trino    ss       27.427           s/op
```

Also, the BaseJavaPruneCompilerSuite has been updated to assert the
invariant that running a subsequent request with changed contents gives
different results (we are not doing incorrect caching).