Cancel stale didChange requests (#227)
Previously, Metals didn't cancel didChange requests. This meant that as
users are typing in a very large file, they could queue up a long list
of requests to typecheck the code for errors. The compiler prioritizes
the most recent requests so it would give users reasonably fast feedback
on their most recent changes, but Metals would eventually process the
old stale requests and potentially publish diagnostics for old versions
of the open file. For Java, we're dealing with performance issues in
some cases as well where these requests can take up to eight seconds to
process, and this queue of `didChange` could become very large. 

Now, we use meaningful cancelation tokens for `didChange` requests, and
we appropriately cancel old requests so that we skip processing them
when they get off the request queue.

<!-- jj-stack-begin -->
- → **#227** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/227/commits/40384ee1234980fdb8527ddafdfc75d8d03826fd)]
- #238
[[Commit](https://github.com/REDACTED_ORG/metals/pull/238/commits/31c67bfa55d14e87d48aeb476c9bbc719f1af3f2)]
- #210
[[Commit](https://github.com/REDACTED_ORG/metals/pull/210/commits/cd59fd5d7b4b66d02eda39eca9478cf31819257b)]
- #239
[[Commit](https://github.com/REDACTED_ORG/metals/pull/239/commits/b6f702ebdbea843e7aa6c0e2dce4241b174a24c8)]
<!-- jj-stack-end -->