Disable prompt to re-import build if `buildChangedAction` is disabled (#241)
I noticed that I have a long-running "Changed" progress task in my
status bar while working on Metals today. This happens because Metals is
waitin on me to reply to the request

> sbt build needs to be re-imported: "Not now" | "Import changes"

Now that we have a `"buildChangedAction": "none"` setting, we should
disable this prompt when that setting is enabled.

## Example

The progress bar on the top is tied to the prompt below it

<img width="354" height="124" alt="image"
src="https://github.com/user-attachments/assets/65b37d15-b8db-4c8c-9616-bc668646dc11"
/>