Correctly resolve cancelled compiler requests (#239)
Previously, if a request to the presentation compiler had been cancelled
before it started, the future would never complete. The `onComplete`
callbacks would never get invoked, and likely leak memory.

Now, instead of throwing an exception in the compiler job queue, we
resolve the request promise with a `CancelationException`, as it's
intended to be done.

<!-- jj-stack-begin -->
- #227
[[Commit](https://github.com/REDACTED_ORG/metals/pull/227/commits/40384ee1234980fdb8527ddafdfc75d8d03826fd)]
- #238
[[Commit](https://github.com/REDACTED_ORG/metals/pull/238/commits/31c67bfa55d14e87d48aeb476c9bbc719f1af3f2)]
- → **#239** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/239/commits/b6f702ebdbea843e7aa6c0e2dce4241b174a24c8)]
<!-- jj-stack-end -->