[PLAT-147717] Support range formatting with Scalafmt (#243)
Previously, there was no way to selectively run Scalafmt on a selected
text range. This was problematic for people working on codebases like
apache/spark that have not enabled Scalafmt on the entire repo but they
want to use Scalafmt for the code they contribue. IntelliJ has supported
range formatting for this use-case, and it's a legitimate feature gap to
staying with IntelliJ over switching to Metals.

Now, there is a new `rangeFormattingProvider` setting that supports
enabling Scalafmt range formatting.
```
"metals.rangeFormattingProvider": ["scalafmt"]
```

When enabled, we run Scalafmt on the *entire file*, and then run Myer's
diff at the token level to extract only the formatted tokens from the
requested range.

Here's a demo of the feature in action:



https://github.com/user-attachments/assets/12b2b6b2-3770-4265-a907-5e96985c3eb1


<!-- jj-stack-begin -->
- #240
[[Commit](https://github.com/REDACTED_ORG/metals/pull/240/commits/8d53a45b8686c817c52be4ceb5d23f8aca862bd3)]
- #242
[[Commit](https://github.com/REDACTED_ORG/metals/pull/242/commits/fd192f1eca49590a592f4a0326b52f1de25fd700)]
- → **#243** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/243/commits/a7e085f6bd4adbf7495d364f7a91adc5fedbe558)]
<!-- jj-stack-end -->