Fix navigation to Java enums with an implicit constructor (#249)
Previously, goto-definition and workspace/symbol didn't work for Java
enums that had the implicit constructor.
```java
enum Day {
  WORKDAY,
  WEEKEND
}
```
The problem was that we didn't get the correct range of the enum
members. This was a regression that was specific to mtags-javac.

Now, we emit the correct location for this (very common) case so that
definition and workspace/symbol work as expected.

## Test Plan

See updated tests. I also manually confirmed that this fixes the issue
where I reproduced it first in the Metals codebase navigating to the
`pc.FeatureFlag` enum.

<!-- jj-stack-begin -->
- → **#249** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/249/commits/1305e68aec7dacaeb1e1611e378b66e1d0f7dda5)]
<!-- jj-stack-end -->