Add repo-wide Find References/Implementations for Java (#210)
Previously, "Find references/implementations" for Java would only work
within Java
files. Now, it works across the entire repo including for unsynced
files.

This functionality is achieved with the new mbt-v2 symbol index, which
indexes both symbol outlines and "fuzzy references", which is
effectively the set of non-local identifiers that are mentioned in each
file. This set of fuzzy references is inserted into the mbt-v2 bloom
filters, just like the `IdentifierIndex` already works for synced files.
The only difference to the existing `IdentifierIndex`, is that we
distinguish between references to

- Overriden methods/types, to power "Find implementation"
- Guaranteed method invocations, to limit false positives. For example,
  if we are searching for usages of the method `execute()`, then we will
  skip files that only reference a *field` named `execute`. For Scala,
  this optimization doesn't work because you can't syntactically
  distinguish between field references and method invocations.

<!-- jj-stack-begin -->
- → **#210** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/210/commits/a03b80c1d6a454ddf0fd3f4d051eac814f1e6b6a)]
<!-- jj-stack-end -->