[PLAT-148404] Hide Incosistent class/module pair errors (#256)
This error seems to pop up from time to time, more often in the runtime
repository. A proper fix is probably difficult and may come later,
almost surely inside the compiler. In the meantime, we downgrade log
level to debug and don't add squiggles that otherwise would pollute the
whole editor and make it unusable.