Go to definition inside binary jars (#247)
When the definition does not have attached sources, use a decompiler to
navigate to its definition.

Previously, Go to definition would fail for external dependencies that
have no attached sources, such as shaded dependencies or protobuf code.

Added cfr dependency to Metals instead of resolving it on the fly and
added infrastructure to make it easy to swap the decompiler to
vineflower or fernflower.

There are still a number of improvements that can land in separate PRs:

- currently, there is no Java editor support in the decompiled
classfile, it would be good to make it work as a regular, read-only,
Java file
- correctly navigate to decompiled Scala code. For instance, the symbol
for an import is usually a term but Java does not have objects, to
top-level classes may sometimes not be correctly located.