Add LSP command for sync-file (#253)
Currently, there's no way to sync files in non-Cursor editors. Alex
Brassel has been using Metals recently with the Helix editor, and
recently documented the setup
https://REDACTED_GITHUB_URL/universe/pull/1434017
So far, he needs to open Cursor to trigger the sync, which is annoying
and only acceptable as a temporary workaround. With this change, the
build sync is now additionally implemented as a standard LSP command so
it can be triggered by any LSP client.

## Test Plan

See the updated ManualSuite. This test case stresses the code path where
a generic LSP client opens the editor and triggers the file-sync
command. I manually confirmed that it works as expected. We don't have
infrastructure for testing file syncing, otherwise.

<!-- jj-stack-begin -->
- → **#253** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/253/commits/8816eb5e6037f46afe3be7470ea385e39e8dbb36)]
<!-- jj-stack-end -->