Bump required version for Scala files to trigger re-index (#270)
Now that we added support for multiple packages per file, we need to
force re-indexing saved documents. We don't have a version for the mbt
index, but we can use the bloom filter version to force re-indexing of
Scala files in order to populate more than one package definition per
file.