No-op: Vendor Turbine Codebase (#269)
This is the first step towards fixing Java performance issues.
So far, we have used the `javac -sourcepath` option to load 1st-party
symbols in the repo. This scales well OK up to codebases of ~100k lines,
but starts to feel sluggish after that. Despite doing optimizations like
pruning method bodies, the Java compiler still eagerly runs the "enter"
phase where it recursively crawls the codebase at the *symbol outline*
level, and this can be expensive, and it needs to run on every single
edit.

Turbine is a Java header compiler that is primarily used in Bazel to
build "ijars", effectively classfiles that only contain signatures (no
method bodies or field initializers). Upstream Turbine is not published
to Maven Central, and it does not gracefully handle broken code making
it unsuitable for a LSP server out of the box. However, with a few
modifications, we have validated that Turbine can recover gracefully
from
errors making it suitable for use in an IDE.

Our benchmarks show that Turbine can build a classpath at 500k-1m lines
of Java code per second during how hot the JVM is. This is fast enough
that we can effectively build the classpath from scratch during every
startup. With some tweaks, we can make this process incremental so it's
lightning fast in most cases.

Given that upstream Turbine is not going to accept diffs for IDE usage,
we vendor the codebase here. Ideally, we could fork the repo and build
it from source via Bazel, but that would require a more complicated
GitHub repo setup and a more complicated build. Instead, we just copy
the code over here even if it means loosing out on the original git
history.

<!-- jj-stack-begin -->
- → **#269** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/269/commits/020596c0bbc4c79c71007b2bb31fcf2bd0a34cdb)]
- #264
[[Commit](https://github.com/REDACTED_ORG/metals/pull/264/commits/58eb1ece95577eafc819de3f82c1fc1501e880a9)]
<!-- jj-stack-end -->