Add mbt as a fallback definition provider (#278)
This is in preparation for the upcoming turbine mode for Java. Since
javac will soon load 1st-party and 3rd-party symbols through the
classpath, it won't be able to implement goto-definition for 1st-party
symbols via the -sourcepath anymore.

Now, when the mbt-based definition provider is enabled, the mbt index
can answer goto-definition requests. This index should work much more
reliably than the old definition index since it does not hold any
inverse index of symbols->files. Instead, we just enumerate all files
that declare symbols in the matching package and filter for matching
definition occurrences.

## Test Plan

None. This is disabled by default, and it there's no use for it until we
add Turbine for Java. I've extracted it to a separate PR so it's cleaner
to read, and we'll automatically turn on this setting when the turbine
mode is enabled so there will in practice only be a single setting.

<!-- jj-stack-begin -->
- → **#278** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/278/commits/3dc0f9aa85fc077197d26a8555a6967cbdad7256)]
- #279
[[Commit](https://github.com/REDACTED_ORG/metals/pull/279/commits/bd41871b3cb19572854462ea72a05731f074b245)]
- #280
[[Commit](https://github.com/REDACTED_ORG/metals/pull/280/commits/a67efcad2e3013cd802e64c91d4f4af9e78d8355)]
- #281
[[Commit](https://github.com/REDACTED_ORG/metals/pull/281/commits/3088fc8bb4358557fa7ef4fa1542ffc07b893617)]
- #283
[[Commit](https://github.com/REDACTED_ORG/metals/pull/283/commits/b14d0d3fbbb455129e725be77444ee5b0372eb05)]
- #264
[[Commit](https://github.com/REDACTED_ORG/metals/pull/264/commits/c64532c62730e04fa4049b249297204be09cc7d0)]
<!-- jj-stack-end -->