Make mbt index work in non-git repos (#279)
I've several times failed to test Metals recently because the mbt index
is unable to run `git ls-files`. For example, I can't use jj workspaces
because it doesn't create a `.git` alias. We also have to run `git
init` for all tests that depend on mbt-v2. There's nothing unique about
git that mbt depends on, it's only using it to run `git ls-files
--stage`, which we can easily replicate  for any repo by just walking
the file tree directly. It's just slower. This PR implements that so
Metals works out of the box in more cases.

<!-- jj-stack-begin -->
- #278
[[Commit](https://github.com/REDACTED_ORG/metals/pull/278/commits/3dc0f9aa85fc077197d26a8555a6967cbdad7256)]
- → **#279** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/279/commits/bd41871b3cb19572854462ea72a05731f074b245)]
- #280
[[Commit](https://github.com/REDACTED_ORG/metals/pull/280/commits/a67efcad2e3013cd802e64c91d4f4af9e78d8355)]
- #281
[[Commit](https://github.com/REDACTED_ORG/metals/pull/281/commits/3088fc8bb4358557fa7ef4fa1542ffc07b893617)]
- #283
[[Commit](https://github.com/REDACTED_ORG/metals/pull/283/commits/b14d0d3fbbb455129e725be77444ee5b0372eb05)]
- #264
[[Commit](https://github.com/REDACTED_ORG/metals/pull/264/commits/c64532c62730e04fa4049b249297204be09cc7d0)]
<!-- jj-stack-end -->