[PLAT-147577] Add main class lenses from all reverse dependencies (#287)
Metals currently assumes that the source file belongs to the same target that runs the test. This is not necessarily the case, since in bazel is common to build in a library target and depend on it from a test target as a runtime dependency. This changes the codelens provider to look for all the main class in all the transitive reverse dependencies since any of those target could run the main.