Navigate to definition via compiler locations (#261)
When a symbol comes from a source file, the presentation compiler knows
it's exact location and it is readily available in sym.pos. Previously,
only symbols coming from the same source file were resolved to sym.pos,
while the other symbols were looked up via indexes. This change picks up
the sym.pos whenever it is available. Note that without this change,
navigation would not work at all for unsynced files, despite the
presentation compiler being able to "understand" most 1st party code
(this is because we don't index sources that are not synced).

The consequences are:
- faster resolution time
- consistent results: if the symbol is resolved by the compiler, it is
going to be found



https://github.com/user-attachments/assets/7231f593-2a7e-4967-a396-6b0c8bea0566