Make Turbine gracefully recover from erroneous code (#281)
Previously, Turbine would fail fast if you ran it on a codebase with
lots of errors. Many of these errors come from missing dependencies, or
because the same symbol is defined in multiple files, or the file has
strange identifiers like `foo<caret>bar()`, etc.

Now, you can feed Turbine a bunch of Java code and it will more or less
always produce a meaningful classfiles for as many symbols as possible.

## Test Plan

I have only manually tested this by running Metals in large repos like
Eclipse JDT, IntelliJ Community, Android AOSP, NetBeans, and Trino.

I think it's OK to go ahead and merge this as long as all the diffs are
correctly commented with `// TURBINE-DIFF` markers.

<!-- jj-stack-begin -->
- #278
[[Commit](https://github.com/REDACTED_ORG/metals/pull/278/commits/3dc0f9aa85fc077197d26a8555a6967cbdad7256)]
- #279
[[Commit](https://github.com/REDACTED_ORG/metals/pull/279/commits/bd41871b3cb19572854462ea72a05731f074b245)]
- #280
[[Commit](https://github.com/REDACTED_ORG/metals/pull/280/commits/a67efcad2e3013cd802e64c91d4f4af9e78d8355)]
- → **#281** ←
[[Commit](https://github.com/REDACTED_ORG/metals/pull/281/commits/3088fc8bb4358557fa7ef4fa1542ffc07b893617)]
- #283
[[Commit](https://github.com/REDACTED_ORG/metals/pull/283/commits/b14d0d3fbbb455129e725be77444ee5b0372eb05)]
- #264
[[Commit](https://github.com/REDACTED_ORG/metals/pull/264/commits/c64532c62730e04fa4049b249297204be09cc7d0)]
<!-- jj-stack-end -->